Split parts up into separate files
