+++ src/qr.rs
@@ -1,0 +1,3 @@
+//! Pivoted QR factorization and specialized solver for systems with
+//! diagonal updates.
+
