+++ src/trust_region.rs
@@ -1,0 +1,1 @@
+//! Solver for the trust-region sub-problem in the LM algorithm.
